+++ Docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimization of table columns in `OpenTgResearcherDesktop`
 - Fixing the Telegram parsing cancellation operation
 - Fixing auto-connection to Telegram when starting `OpenTgResearcherConsole`
+- Fixed Flood Control display when parsing chat in `OpenTgResearcherConsole`
 ### Changed
 - NuGet package update
 - Optimization of `TgStorage`